Address dain's review comments

+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/CombineHashFunction.java
@@ -24,6 +24,7 @@ public final class CombineHashFunction
     @SqlType(StandardTypes.BIGINT)
     public static long getHash(@SqlType(StandardTypes.BIGINT) long previousHashValue, @SqlType(StandardTypes.BIGINT) long value)
     {
-        return 31 * previousHashValue + value;
+        // We use int hashes everywhere, so force this to be int
+        return (int) (31 * previousHashValue + value);
     }
 }
